moved tetrominos.js into utils folder and corrected import statements
